Store results of rule decisions
As rendered configs rely on decisions made in rules and those
potentially be random in nature, for example a dice roll to determine
the bucket of an experiment, we need to store the results in order to
deterministically deliver the same experience to a user. We express this
with a list of integers which could be used to represents dice rolls for
percentage-based rollout, bucket assignment and more unforeseen at this
point.

* introduce ruleDecisions type
* change user repo interface to expect rule decisions
* update postgres implementation to store decisions as json blob