Setup Console
Console will serve as interface for operators of Configsum to setup and
control major aspects of the system. Anticipated are creation of Apps
and their tokens, inspection of rendered configs and most importantly
rule managemnt to setup toggles and experiments.

We decided to go forward with Elm as our frontend technology for its
runtime safety and lack of complex abstractions. The static assets are
compiled into a Go file to avoid the overhead of shipping asset files
along the main binary and spare us distribution headache.

* new subcommand: `console`
* ui file layout
* ui package
* path based routing
* basic styles
* add common operations as make targets