Introduce Clients
To separate configs and tie auth to an entity we introduce clients which
can be understood as distinct consumers of the system (e.g. iOS app, SPA
webapp, or other server components). Clients will have an amount of
tokens to be used for authentication and easy rotation in the event of
a leak.

* add client pkg
* client postgres implementation
* token postgres implementation
* add generate pkg for helper methods