Implement publich token based AuthMiddleware
After introducing Client with their Tokens we want to enable using these
for secret based authentication, where the consumer sends the generated
secret in the `X-Configsum-Token` custom header. In the middleware
inspects the header checks for validity of the secret and lookups the
corresponding Token to find the associated Client. Should one of these
steps fail the middleware rejects the request and no following endpoint
will be called.

* public middleware
* client service
* HTTPToContext for lifting of the secret into the request context
* inmem implementations for client repos
* breaking out repo tests for re-use
* generate method for secure token generation