Implement Dory authentication middleware
Dory is the first auth method we want to support to enable use inside of
Lifesum infrastructure, who will be one of the first users of the
system. It uses a secret based signature check to validate the auth
origin. Testing for this signature should give reasonable confidence
that the information provided in the `X-Dory-Userid` header can be
trusted.

* public middleware
* HTTPToContext for lifting signature & userid into request context