Implement iOS SDK
The iOS sdk represents the means for iOS clients to consume the
Configsum service, allowing client applications to employ remote
configurations, feature toggle. A/B testing.

The entry point is offered by Configsum which exposes the APIs consumers
will interact with.

The user of this sdk is expected to provide an initial configuration comprised
of different parameters for the http/https request as well as an
attributes object that holds device information and user provided
domain specific metadata.

It's logic can be devided in two parts:
* setting up the framework and initial request happens in AppDelegate.
The request retrieves the latest configutation and the sdk will persist
it locally.
* the rest of the APIs will retrieve the needed values based on the
provided keys or return a default value in case it encounters a problem

This sdk does not have third party dependencies as we wanted to avoid
Carthage's cumbersome approach to sub-dependencies