Centralise errors
TO avoid duplication of error infrastructure we introduce a single
package called `errors` which additionally to wrapping the core
functions from github.com/pkg/errors exposing a set of errors which can
be used for control flow through out the system. While an experiment we
remain open to revert this decision and have package local errors, if
those will be consumed outside of Configsum (unforeseen at this point).

* introduce `pkg/errors`
* wrap functions from github.com/pkg/errors
* move all custom errors into central package
* adjust usage accordingly