Simplify metric collection
Turns out all relevant information can be expressed on a single
histogram as those produce also `_count` timeseries which function as
a counter and makes distinct counters obselete if we track the same
dimensions anyway. While reducing the surface this change also
bundles most concerns of instrumentation in the package itself, so that
the cmd code has less knowledge about it.

* bundle metrics collection in histograms
* move metrics related code to package
* update instrumentation implementations