Rules
Rules lie at the core of our sytem. Rules are used by users to alter base
configs by ways of applying new rules and thus rendering the desired
configuration. To start with there are three kinds of rules: global
override, percentage-based rollout and bucketed experiments. The
consumer defines buckets used for distributing client users evenly/unevenly.
The rules are in edit mode until activated at which point they cannot be
modified. The input for rules is a set of values with predefined types
which is enforced at multiple layers in the system.

* implement operations for storing and retrieving rules for postgres
implementation as well as inmem
* implement validation method
* implement method for applying a rule
* define ways to match on the supported parameter types implement tests